The code is made cleaner
